Pre-tested code, with one segment for completion
Need to look at the part in Join, not sure what to do with the child
after we find the first one that quit.

P1_Quit and P1_Kill should be fine. I can't get the makefile to compile
anything on this system so I'm going to move over to my Mac if this
doesn't work, but I wrote out the dispatcher in a way that I THINK
should work.

This should be pretty much it.